prevent default images from being deleted

+++ src/main/java/com/team7/rupiapp/service/UserServiceImpl.java
@@ -101,7 +101,7 @@ public class UserServiceImpl implements UserService {
                 }
 
                 String previousAvatar = user.getAvatar();
-                if (previousAvatar != null) {
+                if (previousAvatar != null && !previousAvatar.equals("default.png")) {
                     Path previousFilePath = uploadPath.resolve(previousAvatar);
                     Files.deleteIfExists(previousFilePath);
                 }
